Minor: Extracted logic from Program.cs to ExchangeRateUpdater (temporary), added Hosting and HTTP to allow for IHttpClientFactory implementation
