+++ jobs/Backend/Task/Program.cs
@@ -1,41 +1,12 @@
-﻿using System;
-using System.Linq;
+﻿using ExchangeRateUpdater.Services;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 
-namespace ExchangeRateUpdater;
 
-public static class Program
-{
-    private static IEnumerable<Currency> currencies = new[]
+var host = new HostBuilder()
+    .ConfigureServices(services =>
     {
-        new Currency("USD"),
-        new Currency("EUR"),
-        new Currency("CZK"),
-        new Currency("JPY"),
-        new Currency("KES"),
-        new Currency("RUB"),
-        new Currency("THB"),
-        new Currency("TRY"),
-        new Currency("XYZ")
-    };
-
-    public static void Main(string[] args)
-    {
-        try
-        {
-            var provider = new ExchangeRateProvider();
-            var rates = provider.GetExchangeRates(currencies);
-
-            Console.WriteLine($"Successfully retrieved {rates.Count()} exchange rates:");
-            foreach (var rate in rates)
-            {
-                Console.WriteLine(rate.ToString());
-            }
-        }
-        catch (Exception e)
-        {
-            Console.WriteLine($"Could not retrieve exchange rates: '{e.Message}'.");
-        }
-
-        Console.ReadLine();
-    }
-}
+        services.AddHttpClient();
+        services.AddTransient<ExchangeRateService>();
+    })
+    .Build();
